Refactor SDK to be more object oriented (#71)
* Update outbound documentation link

* Initial refactoring

* Continued refactoring

* Complete initial refactor

* WIP fixing tests

* Remove extra return

* Continue refactoring

* Fix asset.from_url()

* Clean up example

* Fix iteration problems in python 3.8.6

* fix ClientVersion version staticmethod for py27

* Fix integration test

* get_folder -> get_children

* Fix integration test downloads

* Fix format_headers for py27

* Fix all staticmethods for py27
