+++ frameioclient/__init__.py
@@ -1,2 +1,3 @@
 from .client import FrameioClient
-from .utils import stream
+from .service import *
+from .lib import *
