+++ frameioclient/service/service.py
@@ -1,0 +1,3 @@
+class Service(object):
+    def __init__(self, client):
+        self.client = client
