+++ tests/conftest.py
@@ -4,11 +4,6 @@ from frameioclient import FrameioClient
 
 token = os.getenv('FRAME_IO_TOKEN')
 
-# @pytest.fixture
-# def frameioclient(token):
-#   return FrameioClient("aaaabbbbccccddddeeee")
-
-
 @pytest.fixture()
 def setup_client():
   client = FrameioClient(token)
